refactor

+++ remove_module.go
@@ -18,7 +18,10 @@ func HandleRemoveModule(args []string) {
 		println("Module name is required")
 		return
 	}
-
+	if module == "gerard" {
+		println("Cannot remove the 'gerard'. Are you stupid?")
+		return
+	}
 	err := os.RemoveAll(module)
 	if err != nil {
 		println("Error removing module:", err.Error())
